refactor: migrate auth.asmx/ValidateCredentials to /api/auth/validate
